requirement.txt
discord.py
asyncpraw
python-dotenv
